requirements.txt
# Flask and web framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5

# Database
pymongo==4.5.0

# Authentication and security
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0

# Data validation and serialization
marshmallow==3.20.1
marshmallow-mongoengine==0.31.0

# AI integration (commented out for Vercel compatibility)
# google-generativeai==0.3.1

# HTTP requests and API
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# Date and time utilities
python-dateutil==2.8.2

# Logging and monitoring
Werkzeug==2.3.7

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
factory-boy==3.3.0

# Rate limiting
Flask-Limiter==3.5.0
redis==5.0.1

# API documentation
flasgger==0.9.7.1

# Email support (optional)
Flask-Mail==0.9.1

# Background tasks
celery==5.3.4
